Make some tests pass
