+++ iml-services/iml-devices/src/build_new_state.rs
@@ -8,7 +8,10 @@ use crate::{
     merge_state::merge_state,
     virtual_device::make_other_device_host,
 };
-use iml_wire_types::{db::DeviceId, Fqdn};
+use iml_wire_types::{
+    db::{DeviceId, DeviceType, Device},
+    Fqdn,
+};
 use itertools::Itertools;
 use std::collections::{BTreeMap, BTreeSet};
 
@@ -19,7 +22,7 @@ pub fn build_new_state<'a>(
     db_devices: &Devices,
     db_device_hosts: &DeviceHosts,
 ) -> (Devices, DeviceHosts) {
-    let (temporary_devices, temporary_device_hosts) = merge_state(
+    let (new_devices, temporary_device_hosts) = merge_state(
         fqdn,
         incoming_devices,
         incoming_device_hosts,
@@ -27,7 +30,7 @@ pub fn build_new_state<'a>(
         db_device_hosts,
     );
 
-    let sorted_device_ids: BTreeMap<DeviceId, i16> = temporary_devices
+    let sorted_device_ids: BTreeMap<DeviceId, i16> = new_devices
         .iter()
         .sorted_by_key(|(_, d)| d.max_depth)
         .map(|(id, d)| (id.clone(), d.max_depth))
@@ -39,19 +42,28 @@ pub fn build_new_state<'a>(
             .sorted_by(|((id1, _), _), ((id2, _), _)| {
                 Ord::cmp(&sorted_device_ids[id1], &sorted_device_ids[id2])
             });
-    let sorted_device_hosts2 = sorted_device_hosts.clone();
-    let local_virtual_device_hosts = sorted_device_hosts.filter(|(_, dh)| dh.local);
-    let non_local_virtual_device_hosts = sorted_device_hosts2.filter(|(_, dh)| !dh.local);
+
+    let virtual_device_hosts = sorted_device_hosts.filter(|((id, _), _)| {
+        new_devices
+            .get(id)
+            .map(|d| is_virtual_device(d))
+            .unwrap_or(false)
+    });
+    let virtual_device_hosts_2 = virtual_device_hosts.clone();
+
+    let local_virtual_device_hosts = virtual_device_hosts.filter(|(_, dh)| dh.local);
+    let non_local_virtual_device_hosts = virtual_device_hosts_2.filter(|(_, dh)| !dh.local);
 
     let mut new_device_hosts = temporary_device_hosts.clone();
-    let new_devices = temporary_devices;
 
-    let other_fqdns: BTreeSet<_> = temporary_device_hosts
-        .iter()
-        .filter_map(|((_, f), _)| if f != fqdn { Some(f) } else { None })
-        .collect();
+    for ((_, dh_fqdn), dh) in local_virtual_device_hosts {
+        tracing::info!("Local virtual device host: {}", dh);
 
-    for (_, dh) in local_virtual_device_hosts {
+        let other_fqdns: BTreeSet<_> = temporary_device_hosts
+            .iter()
+            .filter_map(|((_, f), _)| if f != dh_fqdn { Some(f) } else { None })
+            .collect();
+
         for f in other_fqdns.iter() {
             let f = Fqdn(f.0.clone());
             let f2 = f.clone();
@@ -61,18 +73,27 @@ pub fn build_new_state<'a>(
             if all_available {
                 let other_host = make_other_device_host(dh.device_id.clone(), f2, Some(dh));
 
+                tracing::info!("Adding device {:?} to host {:?}", dh.device_id, f3);
                 new_device_hosts.insert((dh.device_id.clone(), f3), other_host);
             }
         }
     }
 
-    for (_, dh) in non_local_virtual_device_hosts {
+    for ((_, dh_fqdn), dh) in non_local_virtual_device_hosts {
+        tracing::info!("Non-local virtual device host: {}", dh);
+
+        let other_fqdns: BTreeSet<_> = temporary_device_hosts
+            .iter()
+            .filter_map(|((_, f), _)| if f != dh_fqdn { Some(f) } else { None })
+            .collect();
+
         for f in other_fqdns.iter() {
             let f = Fqdn(f.0.clone());
             let f3 = f.clone();
             let all_available =
                 are_all_parents_available(&new_devices, &new_device_hosts, f, &dh.device_id);
             if !all_available {
+                tracing::info!("Removing device {:?} from host {:?}", dh.device_id, f3);
                 new_device_hosts.remove(&(dh.device_id.clone(), f3));
             }
         }
@@ -100,6 +121,14 @@ fn are_all_parents_available(
         all_available
     );
     all_available
+}
+
+fn is_virtual_device(device: &Device) -> bool {
+    device.device_type == DeviceType::MdRaid
+        || device.device_type == DeviceType::VolumeGroup
+        || device.device_type == DeviceType::LogicalVolume
+        || device.device_type == DeviceType::Dataset
+        || device.device_type == DeviceType::Zpool
 }
 
 #[cfg(test)]
